requirements.txt
flask 
pytesseract
opencv-python
Pillow
numpy
pandas
pdf2image
langchain==0.2.14
langchain-core==0.2.39
langchain-community==0.2.12
langchain_groq==0.1.9